Simplify span handling, add infrastructure inheritence (#149)
This adds a wrapped span that is used to inherit the infrastructure attribute on spans, it adds a non-WithSpan sequence, modifies users.go & DBUsers.go to use that, with some cleanup of the http error handling.

The new start span, and the existing trace event methods, are moved to a common tracing level.  The client and server subdirectories will move to contain client and service interceptors, as the WithSpan variants are removed.