Backlog of tracing features (#161)
* Backlog of tracing features

This change adds functions to update the span name and reason text.  It also removes the Xxxf tracing variants and enhancing the Xxx forms to do formatting if needed.   And the close flag on the forwarder loop is moved as per a previous review comment.

Finally, users.go has been updated to change the update the span names to something more specific once we know the username.

* apply CR comments