Add blade boot sequence
This adds the blade boot sequence.  This is a delayed effect of powering on, which means that it is triggered by a timer.  That forced a refactoring of the other unit tests.

This also fixes an issue with linked spans that are started while another span is active in the context.

STill to be done: Tor messages are incomplete, get status is missing.  Workload state machine is missing.  GRPC service is missing.