Reduce test key set size to aid test runtime for disk based stores
Currently, a number of the store tests use a set size of 100 to test the function of the stores multi-key read/write/update/delete  calls. There is no real need for this to be so large and reducing this to a more reasonable 10 will significantly ease test runtime for hard disk based stores for no cost in test coverage.