Clean up trace and timestamp client initialization
This is a follow-on from an earlier partial change to handle proper initialization code for these two clients.

They now return an 'already initialized' error that is used in production to indicate an internal error.  Test suites expect to get this, and that it is harmless, so the error is ignored in those cases.

They both cache the grpc connection and clients, replacing them after a failure.  There is a unit test for the timestamp client that verifies this handling.

There is a common pattern and almost perfectly common code between these initializations.  I may factor that out in the future, even though it is a small amount of code.