Clean up test sync issues (#185)
The state machine mechanisms are correctly synchronized with respect to themselves, but the actual state change is placed after the response is issued.  That means that the unit tests, which reach around the normal synchronization, need to be aware that it may take a brief time for the state change to show up.  This was missed outside of the blade unit tests.  This is fixed in this change.

Likewise, the tests now use the constants for the state IDs, rather than in line strings.