Convert stepper to use simple SM
This change converts the stepper service to use the simple-SM structure that was previously used by the inventory services.

The protoactor dependency is removed, along with the tracing and actor state machine.

The simple SM is renamed to SM.

Given that we have a second use, some SM support features were moved from the inventory use to the core, as they were found to be of general use.