repair manager rules engine (#192)
* repair manager rules engine

This contains the initial implementation of the rules engine that the repair manager will use.  It is mostly complete, with three significant exceptions.

First, the interface to the data tables that drive the rules is missing, which impacts the EvaluationContext and Fetch.

Second, the Proposal struct is currently a placeholder.  It is also partly waiting on the datastore, but it also needs the definition of what a repair order will look like.

Third, the error definitions are currently placeholder values.

* apply CR comments