Use generated .cs protobuf files
Modifies the makefile to generate .cs files from a subset of the protobufs.

Modifies the cmdlets csproj to link to them, and to require the protobuf runtime.

The cmdlets are modified to use these new definitions, tests are modified when required.

The temporary data structure definitions, and the Newtonsoft json support are removed.