Add store.DeleteMultiple() to allow for atomic delete of record with index (#213)
* Add atomic delete for record, index pair

Add store.DeleteMultiple() to provide for atomic delete for multiple records
Update inventory definition delete calls to use store.DeleteMultiple()
Add appropriate errors to allow proper reporting of missing index records.

* Add xxx.Delete() calls to a number of the tests cases

Add Delete() calls to many of the tests. Note, this is not full coverage for the Delete() functions, i.e. not much in the way of error path validation..

Co-authored-by: MikeCJohnsonUK <MikeCJohnsonUk@users.noreply.github.com>