Add ability to use anchors and aliases in inventory YAML config files
There is inherently a lot of repetition in the inventory yaml file which can lead to large, hard to read and almost impossible to edit files describing the inventory. By using YAML anchors and aliases this amount of repetition can be dramatically reduced.

Improve inventory readability by using flow notation for the individual port descriptions for both Pdu and Tor definitions.