Avoid test issues with testify Equal() and protobuf atomicMessageInfo
The protobuf generated definitions include a number of internal use only fields, including state.atomicMessageInfo which normally is not a problem. Unfortunately the testify' package [require|assert].Equal() when comparing structures will compare the atomicMessageInfo field which can be arbitrarily set as a result of a JSON encode/decode, or other operation with impact on the data fields of interest. This lead to a bunch of tests failing as a result of comparing a "don't care" field. As a workaround, generate new struct using the contents of the existing struct and use those as compare sources. Not ideal, but works and is safe.

Found/fixed a few cares where wrong functions were being used especially when manipulating pdu, tor and blade details fields.

Separated out manipulation of bootOnPowerOn flage from bootInfo structure.