Include http traces into the trace log file
I've noticed an occasional message on the web_server console saying that the header was updated twice.  No further information, so haven't been able to track it down. This change replaces the default logger (which is what is writing to stderr) with code that uses the tracing subsystem.  At the very least, I'm hoping it will give a better stack trace - as well as context in the overall trace flow...