Moving common definitions out of observer project (#245)
This updates some of the protobuf-derived structures, and is step 2 in a sequence to move the wire protocols into a common area.

The ts-proto plugin does not provide the level of support that we need, although it comes closer than most.  What we're doing here is to use the output of that plugin as a starting point for constructing the typescript we want.

To do that, this change has the main build emit the .ts file, there is a .ts_ref file that it can be compared to (this check is currently manual).  If there is a difference, the custom .tsx file will need to be updated.  Note that the .ts_ref and .tsx files are checked in, while the .ts file is not.

The observer project copies the .tsx files into its tree and uses them.