Add wall clock time to log output (#265)
This adds the real world time to the log events and to the span start and end entries.

This is only formatted by the text formatters used for the log files and the unit test output.  Events are prefixed by the long date/time format.  Spans are prefixed by the start time in long date/time format and the duration.