Apply CR comments, plus
I've updated blade and rack to use the GetXxx forms to access the protobuf fields.

I also added support for WithName to take a format string, and then removed the fmt.Sprintf calls in the WithName calls to simplify the outer code.

I then refactored the loops in inventory.go that involved the WithName changes, as defer inside a loop was causing a warning in the code inspector.

Finally, I ran all changed files through the code formatter.