Add Impact statements to inventoryd operations (#288)
* Add Impact statements to inventoryd operations

Impact claims added to inventory state machine operations
Impact is moved to the span
uplift from child spans is supported in the text output
reason and span name updates are now done via setting attributes to the span
tsx form of the log entries are updated.

* Apply CR comments

* Apply CR comments, plus

I've updated blade and rack to use the GetXxx forms to access the protobuf fields.

I also added support for WithName to take a format string, and then removed the fmt.Sprintf calls in the WithName calls to simplify the outer code.

I then refactored the loops in inventory.go that involved the WithName changes, as defer inside a loop was causing a warning in the code inspector.

Finally, I ran all changed files through the code formatter.

* Fix misplaced withname

I also looked at the running logs to see if there were others.  Didn't spot any.